(feature) Make compatible with vcpkg manifest mode
vcpkg manifest mode means the project will install packages when CMake is invoked during the build step. As it no longer makes sense to install packages ahead of time, then this can be combined with acgetchell/cdt-image to provide everything necessary so that cloning and building a repo works.

Deprecates acgetchell/cdt-image.